timeout decreased to 5s, element finding improved, more control messages
This update includes:

- Maven project redefined to make easiest for the code to find classes
into the project
- Added some control messages to be able to see what is happening in
some steps (I ask you to leave in there for now so if anything fails
would be easiest for us to find where exactly failed)
- Added some code improvements regarding element finding, now I am not
getting the running locking we got before in Slotmine.
- Timeout has been placed at 5 seconds and everything is working right
in server, so I would say to left it in 5 and launch Live 1 Job. If
anything goes wrong I will raise it to 60 as always.
- Decrease a lot the time of execution for the second iteration, now a
global variable stores the Register Element Css once it is adquired in
first iteration, so it is no need to find for it again in 2nd iteration.